hackathon/gmail_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-secret-manager==2.16.4
sqlalchemy==2.0.23
pg8000==1.30.3
cryptography==41.0.5
python-multipart==0.0.6 
pymysql==1.1.1
cloud-sql-python-connector[pymysql]==1.5.0
openai>=1.0.0
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
psycopg2-binary